requirements.txt
# Dependencias para el Extractor de Datos de Contratación Pública
# Instalar con: pip install -r requirements.txt

# Manipulación de datos
pandas>=1.5.0
openpyxl>=3.0.0

# Web scraping y parsing
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests>=2.28.0

# Automatización del navegador
selenium>=4.15.0
webdriver-manager>=4.0.0

# Inteligencia Artificial
google-generativeai>=0.3.0

# Utilidades adicionales
python-dateutil>=2.8.0